More Multiplayer Work
There's now a chat window (open with C), in addition to the multiplayer window (open with TAB). Both can be dragged around.

When you click the box on the bottom of the chat window, you can type in it. Hit enter to send message or escape to cancel.

In theory this allows the start of work on code that syncs more important data.

Worth noting is that currently it's P2P; All players send their data to all other players. This may cause issues later on, we'll see. If it causes issues I should be able to revert it back to the client=>host=>all_clients model fairly easily.