Multiplayer Map Syncing!
I finally got map syncing to work in multiplayer! What I finally realized is that I don't actually need to serialize and send the entire map and all it's objects, because most tiles will be left as the default versions of themselves. So I created a pair of utility functions to deconstruct the map into a string of characters, and then that string gets passed to the client which takes the string and reassembles it into a map.

The best part is that this entire process takes maybe 2 seconds even for a 100x100 tile map! That's more than efficient enough, and we can send the other details as needed in separate packets (entities, player locations, etc).

Great progress!